requirements.txt
Flask
Flask-Cors
pandas
scikit-learn
numpy
openpyxl
waitress
python-dotenv
requests